fixed ai and wave spawning bugs
ai now ignore player if they are dead. cannot spawn waves when treasury is not placed. and made sure scene changer set all needed statics to null since tis the same scene on retry
If you read this it was very fun just wanted to make the final version nice and playable with the retry working fully ;)